Compact suppress clock response

+++ rust/ethercat-rt/src/endpoint/commands.rs
@@ -520,7 +520,7 @@ pub(super) fn handle_stepper_suppress(
         ctx.server.respond(&stepper_suppress_response_frame(
             correlation_id,
             0,
-            monotonic_ns(),
+            monotonic_ns() as u32,
         ));
         return;
     }
@@ -558,7 +558,7 @@ pub(super) fn handle_stepper_suppress(
     ctx.server.respond(&stepper_suppress_response_frame(
         correlation_id,
         result,
-        monotonic_ns(),
+        monotonic_ns() as u32,
     ));
 }
 const ERR_SEED_HOME_STREAMING: i32 = -826;
