+++ rust/ethercat-rt/src/wire.rs
@@ -429,7 +429,7 @@ pub fn set_ff_lead_response_frame(cid: u32, result: i32) -> Vec<u8> {
     control_frame(MessageKind::SetFfLeadResponse, cid, &body)
 }
 
-pub fn stepper_suppress_response_frame(cid: u32, result: i32, effective_clock: u64) -> Vec<u8> {
+pub fn stepper_suppress_response_frame(cid: u32, result: i32, effective_clock: u32) -> Vec<u8> {
     let body = StepperSuppressResponse {
         result,
         effective_clock,
